+++ ZXBox.Core.Tests/GameBoy/GBFileFormatTests.cs
@@ -26,7 +26,7 @@ public class GBFileFormatTests
         List<int> tiles = new List<int>();
         //gb.PC = 0x100; //Entry point
         gb.DoInstructions(1000000000);
-        for (int b = 0x8000; b <= 0x97FF; b++)
+        for (ushort b = 0x8000; b <= 0x97FF; b++)
         {
             tiles.Add(gb.ReadByteFromMemory(b));
         }
